Selection tool is functioning
Selection tool is functioning with no restrictions. The selection box consumes four dynamic effect slots, but is separate from the Render system, because they are always reserved for the selection. This is not a concern with the limit of 256.

The tool operates quite quickly, with very minimal delay, but using local information to rendering the box, and stored information to calculate the overlap.

The next step is to make the boxes objects, which would allow for draging and dropping, as well as better selection.

Further plans include using this as a tool problem to work on, and perfect, the grid system