Healthbars work!
Created clean, functioning health bars. They adapt to the health the player has and increase size/change style accordingly. The healthbars used for the HUD are also right justified. The player's name is displayed below their icon. The entire lower left hud vanishes when the camera teleports equal to the amount of time the camera takes to move (roughly).